SYM-39: Tests passing for assertions inside action class.
